[ADD/#125] "설정 > 알림설정" 컴포넌트 분리
