+++ app/src/main/java/com/sopt/clody/presentation/ui/setting/component/NotificationSettingTime.kt
@@ -1,0 +1,52 @@
+package com.sopt.clody.presentation.ui.setting.component
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.sopt.clody.R
+import com.sopt.clody.ui.theme.ClodyTheme
+
+@Composable
+fun NotificationSettingTime(showBottomSheetStateChange: (Boolean) -> Unit) {
+    var selectedTime by remember { mutableStateOf("오후 9시 30분") }
+
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentHeight()
+            .padding(horizontal = 24.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            text = stringResource(R.string.notification_setting_notification_time),
+            style = ClodyTheme.typography.body1Medium,
+            color = ClodyTheme.colors.gray03
+        )
+        Spacer(modifier = Modifier.weight(1f))
+        Text(
+            text = selectedTime,
+            style = ClodyTheme.typography.body3Medium,
+            color = ClodyTheme.colors.gray05,
+        )
+        Image(
+            painterResource(id = R.drawable.ic_setting_next),
+            contentDescription = "pick the alarm time",
+            modifier = Modifier.clickable { showBottomSheetStateChange(true) }
+        )
+    }
+}
