+++ app/src/main/java/com/sopt/clody/presentation/ui/setting/screen/NotificationSettingScreen.kt
@@ -10,11 +10,8 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Switch
-import androidx.compose.material3.SwitchDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -27,7 +24,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.sopt.clody.R
 import com.sopt.clody.presentation.ui.component.popup.ClodyPopupBottomSheet
-import com.sopt.clody.presentation.ui.setting.component.SettingBottomSheetTimePicker
+import com.sopt.clody.presentation.ui.setting.component.NotificationSettingSwitch
+import com.sopt.clody.presentation.ui.setting.component.NotificationSettingTime
+import com.sopt.clody.presentation.ui.setting.component.NotificationSettingTimePicker
 import com.sopt.clody.presentation.ui.setting.component.SettingTopAppBar
 import com.sopt.clody.presentation.ui.setting.navigation.SettingNavigator
 import com.sopt.clody.ui.theme.ClodyTheme
@@ -61,7 +60,7 @@ fun NotificationSettingScreen(onBackClick: () -> Unit) {
                 checkedState = remember { mutableStateOf(true) }
             )
             Spacer(modifier = Modifier.height(32.dp))
-            AlarmTimeSetting(
+            NotificationSettingTime(
                 showBottomSheetStateChange = { newState -> showNotificationTimePicker = newState }
             )
             Spacer(modifier = Modifier.height(32.dp))
@@ -73,7 +72,7 @@ fun NotificationSettingScreen(onBackClick: () -> Unit) {
     }
     if (showNotificationTimePicker) {
         ClodyPopupBottomSheet(onDismissRequest = { showNotificationTimePicker = false }) {
-            SettingBottomSheetTimePicker(
+            NotificationSettingTimePicker(
                 onDismissRequest = { showNotificationTimePicker = false },
                 onTimeSelected = { newTime ->
                     selectedTime = newTime
@@ -81,64 +80,6 @@ fun NotificationSettingScreen(onBackClick: () -> Unit) {
                 }
             )
         }
-    }
-}
-
-@Composable
-fun NotificationSettingSwitch(title: String, checkedState: MutableState<Boolean>) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 24.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = title,
-            style = ClodyTheme.typography.body1Medium,
-            color = ClodyTheme.colors.gray03
-        )
-        Spacer(modifier = Modifier.weight(1f))
-        Switch(
-            checked = checkedState.value,
-            onCheckedChange = { checkedState.value = it },
-            colors = SwitchDefaults.colors(
-                checkedThumbColor = ClodyTheme.colors.white,
-                checkedTrackColor = ClodyTheme.colors.darkYellow.copy(alpha = 0.5f),
-                uncheckedThumbColor = ClodyTheme.colors.white,
-                uncheckedTrackColor = ClodyTheme.colors.gray06,
-                uncheckedBorderColor = ClodyTheme.colors.gray06
-            )
-        )
-    }
-}
-
-@Composable
-fun AlarmTimeSetting(showBottomSheetStateChange: (Boolean) -> Unit) {
-    var selectedTime by remember { mutableStateOf("오후 9시 30분") }
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .wrapContentHeight()
-            .padding(horizontal = 24.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = stringResource(R.string.notification_setting_notification_time),
-            style = ClodyTheme.typography.body1Medium,
-            color = ClodyTheme.colors.gray03
-        )
-        Spacer(modifier = Modifier.weight(1f))
-        Text(
-            text = selectedTime,
-            style = ClodyTheme.typography.body3Medium,
-            color = ClodyTheme.colors.gray05,
-        )
-        Image(
-            painterResource(id = R.drawable.ic_setting_next),
-            contentDescription = "pick the alarm time",
-            modifier = Modifier.clickable { showBottomSheetStateChange(true) }
-        )
     }
 }
